Finished
This is the final product of the dice game. It is called the "Pig Game". I could add rules that pop up at the beginning so the user knows how to play. That is a thought. There are rules in comment form in the JS file. I should have made pushes/merges along the way. All of the buttons work appropriately with the game. The game resets to what it loads to original format when new game button is pressed. JS file is full of comments.