Two Sixes Bug Fixes
I needed to set the scores array for the active player back to 0. Otherwise, behind the scenes the score was still being tallied even if the UI content was updated. I also fixed the fact that if a player pressed hold after rolling a 6, it would make the other player lose all of their points if they rolled a 6, even though they did not roll consecutive sixes on their turn.